Added files via upload
Scrum 2. I have made the paddle move and I made a loop for the bricks. I am drawing the same brick 4 times for example, so that when you hit it, one block is destroyed and 3 more are left (giving the illusion that it takes multiple shots to destroy the block). I also organized the code.